requirements.txt
streamlit
pandas
numpy
plotly
rupture